Add a asynchronous API for performing requests in batches.
The new vectorised functions (_v) submit requests in parallel for
potentially significant performance gains. The precise gains
depend heavily on the type of request (block or scatter) and on
the backend.

Both the Linux/Juju and Mac OS X/IOKit stacks have serious bugs
with regards to how they handle asynchronous requests.  These bugs
have been traced to the Linux/Mach kernels and often trigger
kernel panics. For this reason asynchronous requests are currently
disabled for Linux/Juju and only partially enabled for Mac OS
X/IOKit. It is expected that the bugs in the Linux kernel will be
fixed in the near future. However, as the vectorised APIs have a
lower overhead performance improvements can still be realised.

The biggest gains are found in the Python bindings. By using the
new read API it is possible to contain block read performance on
par with the native C API and scatter performance within 10% of
native code. Further gains can be expected as the quality of
backends improves.

While API/ABI compatibility is preserved the library minor version
is bumped to 0.2.0.