forensic1394_get_nodeid => forensic1394_get_node_id.
This is an API and ABI incompatible change! Programs that make
use of the node ID will need minor modifications. The library
version has been bumped up to 0.3.0 and the SOVERSION to 2.

The function was renamed in order to improve the consistency of
the API c.f, forensic1394_get_vendor_id.